steam::generated

Enum EContentReportReason

#[repr(i32)]
pub enum EContentReportReason {

    KEContentReportReasonInvalid = 0,
    KEContentReportReasonNone = 1,
    KEContentReportReasonUnknown = 2,
    KEContentReportReasonHarassment = 3,
    KEContentReportReasonBullyingAndIntimidation = 4,
    KEContentReportReasonStalking = 5,
    KEContentReportReasonDoxxing = 6,
    KEContentReportReasonOtherHarassment = 7,
    KEContentReportReasonEncouragingViolence = 8,
    KEContentReportReasonEncouragingSelfHarm = 9,
    KEContentReportReasonEncouragingSuicide = 10,
    KEContentReportReasonOtherViolenceOrSelfHarm = 11,
    KEContentReportReasonPhishingOrAccountTheft = 12,
    KEContentReportReasonAttemptedScamming = 13,
    KEContentReportReasonLinkingToMaliciousContent = 14,
    KEContentReportReasonImpersonation = 15,
    KEContentReportReasonOtherScamsAndTheft = 16,
    KEContentReportReasonEncouragingTerrorism = 17,
    KEContentReportReasonOrganizingTerrorism = 18,
    KEContentReportReasonOtherTerrorism = 19,
    KEContentReportReasonTargetedAbuse = 20,
    KEContentReportReasonNamingAndShaming = 21,
    KEContentReportReasonDiscrimination = 22,
    KEContentReportReasonOtherAbuse = 23,
    KEContentReportReasonTrolling = 24,
    KEContentReportReasonBaiting = 25,
    KEContentReportReasonDerailing = 26,
    KEContentReportReasonOtherDisruptive = 27,
    KEContentReportReasonSpam = 28,
    KEContentReportReasonBegging = 29,
    KEContentReportReasonReposting = 30,
    KEContentReportReasonOtherOffTopic = 31,
    KEContentReportReasonCsamSexualContent = 32,
    KEContentReportReasonCsamGroomingOrEnticement = 33,
    KEContentReportReasonCsamOther = 34,
    KEContentReportReasonNudityOrSexualContent = 35,
    KEContentReportReasonNonConsensualMaterial = 36,
    KEContentReportReasonAdvertising = 37,
    KEContentReportReasonReferralLinks = 38,
    KEContentReportReasonGambling = 39,
    KEContentReportReasonRaffles = 40,
    KEContentReportReasonOtherCommercialActivity = 41,
    KEContentReportReasonInauthenticReview = 42,
    KEContentReportReasonHiddenAdvertisementOrCommercialCommunication = 43,
    KEContentReportReasonMisleadingInformationAboutGoodsOrServices = 44,
    KEContentReportReasonMisleadingInformationAboutConsumerRights = 45,
    KEContentReportReasonNoncomplianceWithPricingRegulations = 46,
    KEContentReportReasonRightToBeForgottenViolation = 47,
    KEContentReportReasonMissingProcessingGroundForData = 48,
    KEContentReportReasonOtherDataProtectionAndPrivacyViolation = 49,
    KEContentReportReasonGenderedHarassment = 50,
    KEContentReportReasonGenderedBullyingAndIntimidation = 51,
    KEContentReportReasonGenderedStalking = 52,
    KEContentReportReasonGenderedDoxxing = 53,
    KEContentReportReasonGenderedOtherHarassment = 54,
    KEContentReportReasonGenderedEncouragingViolence = 55,
    KEContentReportReasonGenderedTargetedAbuse = 56,
    KEContentReportReasonCsamFakedSexualContent = 57,
    KEContentReportReasonGenderedNonConsensualMaterial = 58,
    KEContentReportReasonFakedGenderedNonConsensualMaterial = 59,
    KEContentReportReasonFakedNonConsensualMaterial = 60,
    KEContentReportReasonNegativeEffectonDiscourseOrElections = 61,
    KEContentReportReasonQuotesModeratedContent = 62,
    KEContentReportReasonCredibleThreatOfViolence = 63,
    KEContentReportReasonMax = 64,

}

Variants

KEContentReportReasonInvalid = 0

KEContentReportReasonNone = 1

KEContentReportReasonUnknown = 2

KEContentReportReasonHarassment = 3

KEContentReportReasonBullyingAndIntimidation = 4

KEContentReportReasonStalking = 5

KEContentReportReasonDoxxing = 6

KEContentReportReasonOtherHarassment = 7

KEContentReportReasonEncouragingViolence = 8

KEContentReportReasonEncouragingSelfHarm = 9

KEContentReportReasonEncouragingSuicide = 10

KEContentReportReasonOtherViolenceOrSelfHarm = 11

KEContentReportReasonPhishingOrAccountTheft = 12

KEContentReportReasonAttemptedScamming = 13

KEContentReportReasonLinkingToMaliciousContent = 14

KEContentReportReasonImpersonation = 15

KEContentReportReasonOtherScamsAndTheft = 16

KEContentReportReasonEncouragingTerrorism = 17

KEContentReportReasonOrganizingTerrorism = 18

KEContentReportReasonOtherTerrorism = 19

KEContentReportReasonTargetedAbuse = 20

KEContentReportReasonNamingAndShaming = 21

KEContentReportReasonDiscrimination = 22

KEContentReportReasonOtherAbuse = 23

KEContentReportReasonTrolling = 24

KEContentReportReasonBaiting = 25

KEContentReportReasonDerailing = 26

KEContentReportReasonOtherDisruptive = 27

KEContentReportReasonSpam = 28

KEContentReportReasonBegging = 29

KEContentReportReasonReposting = 30

KEContentReportReasonOtherOffTopic = 31

KEContentReportReasonCsamSexualContent = 32

KEContentReportReasonCsamGroomingOrEnticement = 33

KEContentReportReasonCsamOther = 34

KEContentReportReasonNudityOrSexualContent = 35

KEContentReportReasonNonConsensualMaterial = 36

KEContentReportReasonAdvertising = 37

KEContentReportReasonReferralLinks = 38

KEContentReportReasonGambling = 39

KEContentReportReasonRaffles = 40

KEContentReportReasonOtherCommercialActivity = 41

KEContentReportReasonInauthenticReview = 42

KEContentReportReasonHiddenAdvertisementOrCommercialCommunication = 43

KEContentReportReasonMisleadingInformationAboutGoodsOrServices = 44

KEContentReportReasonMisleadingInformationAboutConsumerRights = 45

KEContentReportReasonNoncomplianceWithPricingRegulations = 46

KEContentReportReasonRightToBeForgottenViolation = 47

KEContentReportReasonMissingProcessingGroundForData = 48

KEContentReportReasonOtherDataProtectionAndPrivacyViolation = 49

KEContentReportReasonGenderedHarassment = 50

KEContentReportReasonGenderedBullyingAndIntimidation = 51

KEContentReportReasonGenderedStalking = 52

KEContentReportReasonGenderedDoxxing = 53

KEContentReportReasonGenderedOtherHarassment = 54

KEContentReportReasonGenderedEncouragingViolence = 55

KEContentReportReasonGenderedTargetedAbuse = 56

KEContentReportReasonCsamFakedSexualContent = 57

KEContentReportReasonGenderedNonConsensualMaterial = 58

KEContentReportReasonFakedGenderedNonConsensualMaterial = 59

KEContentReportReasonFakedNonConsensualMaterial = 60

KEContentReportReasonNegativeEffectonDiscourseOrElections = 61

KEContentReportReasonQuotesModeratedContent = 62

KEContentReportReasonCredibleThreatOfViolence = 63

KEContentReportReasonMax = 64

Implementations

impl EContentReportReason

pub fn is_valid(value: i32) -> bool

Returns true if value is a variant of EContentReportReason.

pub fn from_i32(value: i32) -> Option<EContentReportReason>

👎Deprecated: Use the TryFrom<i32> implementation instead

Converts an i32 to a EContentReportReason, or None if value is not a valid variant.

impl EContentReportReason

pub fn as_str_name(&self) -> &'static str

String value of the enum field names used in the ProtoBuf definition.

The values are not transformed in any way and thus are considered stable (if the ProtoBuf definition does not change) and safe for programmatic use.

pub fn from_str_name(value: &str) -> Option<Self>

Creates an enum from field names used in the ProtoBuf definition.