steam::generated

Enum EProtoClanEventType

#[repr(i32)]
pub enum EProtoClanEventType {

    KEClanOtherEvent = 1,
    KEClanGameEvent = 2,
    KEClanPartyEvent = 3,
    KEClanMeetingEvent = 4,
    KEClanSpecialCauseEvent = 5,
    KEClanMusicAndArtsEvent = 6,
    KEClanSportsEvent = 7,
    KEClanTripEvent = 8,
    KEClanChatEvent = 9,
    KEClanGameReleaseEvent = 10,
    KEClanBroadcastEvent = 11,
    KEClanSmallUpdateEvent = 12,
    KEClanPreAnnounceMajorUpdateEvent = 13,
    KEClanMajorUpdateEvent = 14,
    KEClanDlcReleaseEvent = 15,
    KEClanFutureReleaseEvent = 16,
    KEClanESportTournamentStreamEvent = 17,
    KEClanDevStreamEvent = 18,
    KEClanFamousStreamEvent = 19,
    KEClanGameSalesEvent = 20,
    KEClanGameItemSalesEvent = 21,
    KEClanInGameBonusXpEvent = 22,
    KEClanInGameLootEvent = 23,
    KEClanInGamePerksEvent = 24,
    KEClanInGameChallengeEvent = 25,
    KEClanInGameContestEvent = 26,
    KEClanIrlEvent = 27,
    KEClanNewsEvent = 28,
    KEClanBetaReleaseEvent = 29,
    KEClanInGameContentReleaseEvent = 30,
    KEClanFreeTrial = 31,
    KEClanSeasonRelease = 32,
    KEClanSeasonUpdate = 33,
    KEClanCrosspostEvent = 34,
    KEClanInGameEventGeneral = 35,
    KEClanCreatorHome = 36,

}

Variants

KEClanOtherEvent = 1

KEClanGameEvent = 2

KEClanPartyEvent = 3

KEClanMeetingEvent = 4

KEClanSpecialCauseEvent = 5

KEClanMusicAndArtsEvent = 6

KEClanSportsEvent = 7

KEClanTripEvent = 8

KEClanChatEvent = 9

KEClanGameReleaseEvent = 10

KEClanBroadcastEvent = 11

KEClanSmallUpdateEvent = 12

KEClanPreAnnounceMajorUpdateEvent = 13

KEClanMajorUpdateEvent = 14

KEClanDlcReleaseEvent = 15

KEClanFutureReleaseEvent = 16

KEClanESportTournamentStreamEvent = 17

KEClanDevStreamEvent = 18

KEClanFamousStreamEvent = 19

KEClanGameSalesEvent = 20

KEClanGameItemSalesEvent = 21

KEClanInGameBonusXpEvent = 22

KEClanInGameLootEvent = 23

KEClanInGamePerksEvent = 24

KEClanInGameChallengeEvent = 25

KEClanInGameContestEvent = 26

KEClanIrlEvent = 27

KEClanNewsEvent = 28

KEClanBetaReleaseEvent = 29

KEClanInGameContentReleaseEvent = 30

KEClanFreeTrial = 31

KEClanSeasonRelease = 32

KEClanSeasonUpdate = 33

KEClanCrosspostEvent = 34

KEClanInGameEventGeneral = 35

KEClanCreatorHome = 36

Implementations

impl EProtoClanEventType

pub fn is_valid(value: i32) -> bool

Returns true if value is a variant of EProtoClanEventType.

pub fn from_i32(value: i32) -> Option<EProtoClanEventType>

👎Deprecated: Use the TryFrom<i32> implementation instead

Converts an i32 to a EProtoClanEventType, or None if value is not a valid variant.

impl EProtoClanEventType

pub fn as_str_name(&self) -> &'static str

String value of the enum field names used in the ProtoBuf definition.

The values are not transformed in any way and thus are considered stable (if the ProtoBuf definition does not change) and safe for programmatic use.

pub fn from_str_name(value: &str) -> Option<Self>

Creates an enum from field names used in the ProtoBuf definition.