steam::generated

Struct CClanEventData

pub struct CClanEventData {

    pub gid: Option<u64>,
    pub clan_steamid: Option<u64>,
    pub event_name: Option<String>,
    pub event_type: Option<i32>,
    pub appid: Option<u32>,
    pub server_address: Option<String>,
    pub server_password: Option<String>,
    pub rtime32_start_time: Option<u32>,
    pub rtime32_end_time: Option<u32>,
    pub comment_count: Option<i32>,
    pub creator_steamid: Option<u64>,
    pub last_update_steamid: Option<u64>,
    pub event_notes: Option<String>,
    pub jsondata: Option<String>,
    pub announcement_body: Option<CCommunityClanAnnouncementInfo>,
    pub published: Option<bool>,
    pub hidden: Option<bool>,
    pub rtime32_visibility_start: Option<u32>,
    pub rtime32_visibility_end: Option<u32>,
    pub broadcaster_accountid: Option<u32>,
    pub follower_count: Option<u32>,
    pub ignore_count: Option<u32>,
    pub forum_topic_id: Option<u64>,
    pub rtime32_last_modified: Option<u32>,
    pub news_post_gid: Option<u64>,
    pub rtime_mod_reviewed: Option<u32>,
    pub featured_app_tagid: Option<u32>,
    pub referenced_appids: Vec<u32>,
    pub build_id: Option<u32>,
    pub build_branch: Option<String>,
    pub unlisted: Option<bool>,

}

Fields
gid: Option<u64>clan_steamid: Option<u64>event_name: Option<String>event_type: Option<i32>appid: Option<u32>server_address: Option<String>server_password: Option<String>rtime32_start_time: Option<u32>rtime32_end_time: Option<u32>comment_count: Option<i32>creator_steamid: Option<u64>last_update_steamid: Option<u64>event_notes: Option<String>jsondata: Option<String>announcement_body: Option<CCommunityClanAnnouncementInfo>published: Option<bool>hidden: Option<bool>rtime32_visibility_start: Option<u32>rtime32_visibility_end: Option<u32>broadcaster_accountid: Option<u32>follower_count: Option<u32>ignore_count: Option<u32>forum_topic_id: Option<u64>rtime32_last_modified: Option<u32>news_post_gid: Option<u64>rtime_mod_reviewed: Option<u32>featured_app_tagid: Option<u32>referenced_appids: Vec<u32>build_id: Option<u32>build_branch: Option<String>unlisted: Option<bool>
Implementations

impl CClanEventData

pub fn gid(&self) -> u64

Returns the value of gid, or the default value if gid is unset.

pub fn clan_steamid(&self) -> u64

Returns the value of clan_steamid, or the default value if clan_steamid is unset.

pub fn event_name(&self) -> &str

Returns the value of event_name, or the default value if event_name is unset.

pub fn event_type(&self) -> EProtoClanEventType

Returns the enum value of event_type, or the default if the field is unset or set to an invalid enum value.

pub fn set_event_type(&mut self, value: EProtoClanEventType)

Sets event_type to the provided enum value.

pub fn appid(&self) -> u32

Returns the value of appid, or the default value if appid is unset.

pub fn server_address(&self) -> &str

Returns the value of server_address, or the default value if server_address is unset.

pub fn server_password(&self) -> &str

Returns the value of server_password, or the default value if server_password is unset.

pub fn rtime32_start_time(&self) -> u32

Returns the value of rtime32_start_time, or the default value if rtime32_start_time is unset.

pub fn rtime32_end_time(&self) -> u32

Returns the value of rtime32_end_time, or the default value if rtime32_end_time is unset.

pub fn comment_count(&self) -> i32

Returns the value of comment_count, or the default value if comment_count is unset.

pub fn creator_steamid(&self) -> u64

Returns the value of creator_steamid, or the default value if creator_steamid is unset.

pub fn last_update_steamid(&self) -> u64

Returns the value of last_update_steamid, or the default value if last_update_steamid is unset.

pub fn event_notes(&self) -> &str

Returns the value of event_notes, or the default value if event_notes is unset.

pub fn jsondata(&self) -> &str

Returns the value of jsondata, or the default value if jsondata is unset.

pub fn published(&self) -> bool

Returns the value of published, or the default value if published is unset.

pub fn hidden(&self) -> bool

Returns the value of hidden, or the default value if hidden is unset.

pub fn rtime32_visibility_start(&self) -> u32

Returns the value of rtime32_visibility_start, or the default value if rtime32_visibility_start is unset.

pub fn rtime32_visibility_end(&self) -> u32

Returns the value of rtime32_visibility_end, or the default value if rtime32_visibility_end is unset.

pub fn broadcaster_accountid(&self) -> u32

Returns the value of broadcaster_accountid, or the default value if broadcaster_accountid is unset.

pub fn follower_count(&self) -> u32

Returns the value of follower_count, or the default value if follower_count is unset.

pub fn ignore_count(&self) -> u32

Returns the value of ignore_count, or the default value if ignore_count is unset.

pub fn forum_topic_id(&self) -> u64

Returns the value of forum_topic_id, or the default value if forum_topic_id is unset.

pub fn rtime32_last_modified(&self) -> u32

Returns the value of rtime32_last_modified, or the default value if rtime32_last_modified is unset.

pub fn news_post_gid(&self) -> u64

Returns the value of news_post_gid, or the default value if news_post_gid is unset.

pub fn rtime_mod_reviewed(&self) -> u32

Returns the value of rtime_mod_reviewed, or the default value if rtime_mod_reviewed is unset.

pub fn featured_app_tagid(&self) -> u32

Returns the value of featured_app_tagid, or the default value if featured_app_tagid is unset.

pub fn build_id(&self) -> u32

Returns the value of build_id, or the default value if build_id is unset.

pub fn build_branch(&self) -> &str

Returns the value of build_branch, or the default value if build_branch is unset.

pub fn unlisted(&self) -> bool

Returns the value of unlisted, or the default value if unlisted is unset.