steam::generated

Struct CMsgAppRights

pub struct CMsgAppRights {

    pub edit_info: Option<bool>,
    pub publish: Option<bool>,
    pub view_error_data: Option<bool>,
    pub download: Option<bool>,
    pub upload_cdkeys: Option<bool>,
    pub generate_cdkeys: Option<bool>,
    pub view_financials: Option<bool>,
    pub manage_ceg: Option<bool>,
    pub manage_signing: Option<bool>,
    pub manage_cdkeys: Option<bool>,
    pub edit_marketing: Option<bool>,
    pub economy_support: Option<bool>,
    pub economy_support_supervisor: Option<bool>,
    pub manage_pricing: Option<bool>,
    pub broadcast_live: Option<bool>,
    pub view_marketing_traffic: Option<bool>,
    pub edit_store_display_content: Option<bool>,

}

Fields
edit_info: Option<bool>publish: Option<bool>view_error_data: Option<bool>download: Option<bool>upload_cdkeys: Option<bool>generate_cdkeys: Option<bool>view_financials: Option<bool>manage_ceg: Option<bool>manage_signing: Option<bool>manage_cdkeys: Option<bool>edit_marketing: Option<bool>economy_support: Option<bool>economy_support_supervisor: Option<bool>manage_pricing: Option<bool>broadcast_live: Option<bool>view_marketing_traffic: Option<bool>edit_store_display_content: Option<bool>
Implementations

impl CMsgAppRights

pub fn edit_info(&self) -> bool

Returns the value of edit_info, or the default value if edit_info is unset.

pub fn publish(&self) -> bool

Returns the value of publish, or the default value if publish is unset.

pub fn view_error_data(&self) -> bool

Returns the value of view_error_data, or the default value if view_error_data is unset.

pub fn download(&self) -> bool

Returns the value of download, or the default value if download is unset.

pub fn upload_cdkeys(&self) -> bool

Returns the value of upload_cdkeys, or the default value if upload_cdkeys is unset.

pub fn generate_cdkeys(&self) -> bool

Returns the value of generate_cdkeys, or the default value if generate_cdkeys is unset.

pub fn view_financials(&self) -> bool

Returns the value of view_financials, or the default value if view_financials is unset.

pub fn manage_ceg(&self) -> bool

Returns the value of manage_ceg, or the default value if manage_ceg is unset.

pub fn manage_signing(&self) -> bool

Returns the value of manage_signing, or the default value if manage_signing is unset.

pub fn manage_cdkeys(&self) -> bool

Returns the value of manage_cdkeys, or the default value if manage_cdkeys is unset.

pub fn edit_marketing(&self) -> bool

Returns the value of edit_marketing, or the default value if edit_marketing is unset.

pub fn economy_support(&self) -> bool

Returns the value of economy_support, or the default value if economy_support is unset.

pub fn economy_support_supervisor(&self) -> bool

Returns the value of economy_support_supervisor, or the default value if economy_support_supervisor is unset.

pub fn manage_pricing(&self) -> bool

Returns the value of manage_pricing, or the default value if manage_pricing is unset.

pub fn broadcast_live(&self) -> bool

Returns the value of broadcast_live, or the default value if broadcast_live is unset.

pub fn view_marketing_traffic(&self) -> bool

Returns the value of view_marketing_traffic, or the default value if view_marketing_traffic is unset.

pub fn edit_store_display_content(&self) -> bool

Returns the value of edit_store_display_content, or the default value if edit_store_display_content is unset.