steam::generated

Struct CMsgProtoBufHeader

pub struct CMsgProtoBufHeader {

    pub steamid: Option<u64>,
    pub client_sessionid: Option<i32>,
    pub routing_appid: Option<u32>,
    pub jobid_source: Option<u64>,
    pub jobid_target: Option<u64>,
    pub target_job_name: Option<String>,
    pub seq_num: Option<i32>,
    pub eresult: Option<i32>,
    pub error_message: Option<String>,
    pub auth_account_flags: Option<u32>,
    pub token_source: Option<u32>,
    pub admin_spoofing_user: Option<bool>,
    pub transport_error: Option<i32>,
    pub messageid: Option<u64>,
    pub publisher_group_id: Option<u32>,
    pub sysid: Option<u32>,
    pub webapi_key_id: Option<u32>,
    pub is_from_external_source: Option<bool>,
    pub forward_to_sysid: Vec<u32>,
    pub cm_sysid: Option<u32>,
    pub launcher_type: Option<u32>,
    pub realm: Option<u32>,
    pub timeout_ms: Option<i32>,
    pub debug_source: Option<String>,
    pub debug_source_string_index: Option<u32>,
    pub token_id: Option<u64>,
    pub routing_gc: Option<CMsgGcRoutingProtoBufHeader>,
    pub session_disposition: Option<i32>,
    pub wg_token: Option<String>,
    pub webui_auth_key: Option<String>,
    pub exclude_client_sessionids: Vec<i32>,
    pub admin_request_spoofing_steamid: Option<u64>,
    pub is_valveds: Option<bool>,
    pub trace_tag: Option<u64>,
    pub ip_addr: Option<IpAddr>,

}

Fields
steamid: Option<u64>client_sessionid: Option<i32>routing_appid: Option<u32>jobid_source: Option<u64>jobid_target: Option<u64>target_job_name: Option<String>seq_num: Option<i32>eresult: Option<i32>error_message: Option<String>auth_account_flags: Option<u32>token_source: Option<u32>admin_spoofing_user: Option<bool>transport_error: Option<i32>messageid: Option<u64>publisher_group_id: Option<u32>sysid: Option<u32>webapi_key_id: Option<u32>is_from_external_source: Option<bool>forward_to_sysid: Vec<u32>cm_sysid: Option<u32>launcher_type: Option<u32>realm: Option<u32>timeout_ms: Option<i32>debug_source: Option<String>debug_source_string_index: Option<u32>token_id: Option<u64>routing_gc: Option<CMsgGcRoutingProtoBufHeader>session_disposition: Option<i32>wg_token: Option<String>webui_auth_key: Option<String>exclude_client_sessionids: Vec<i32>admin_request_spoofing_steamid: Option<u64>is_valveds: Option<bool>trace_tag: Option<u64>ip_addr: Option<IpAddr>
Implementations

impl CMsgProtoBufHeader

pub fn steamid(&self) -> u64

Returns the value of steamid, or the default value if steamid is unset.

pub fn client_sessionid(&self) -> i32

Returns the value of client_sessionid, or the default value if client_sessionid is unset.

pub fn routing_appid(&self) -> u32

Returns the value of routing_appid, or the default value if routing_appid is unset.

pub fn jobid_source(&self) -> u64

Returns the value of jobid_source, or the default value if jobid_source is unset.

pub fn jobid_target(&self) -> u64

Returns the value of jobid_target, or the default value if jobid_target is unset.

pub fn target_job_name(&self) -> &str

Returns the value of target_job_name, or the default value if target_job_name is unset.

pub fn eresult(&self) -> i32

Returns the value of eresult, or the default value if eresult is unset.

pub fn error_message(&self) -> &str

Returns the value of error_message, or the default value if error_message is unset.

pub fn auth_account_flags(&self) -> u32

Returns the value of auth_account_flags, or the default value if auth_account_flags is unset.

pub fn transport_error(&self) -> i32

Returns the value of transport_error, or the default value if transport_error is unset.

pub fn messageid(&self) -> u64

Returns the value of messageid, or the default value if messageid is unset.

pub fn publisher_group_id(&self) -> u32

Returns the value of publisher_group_id, or the default value if publisher_group_id is unset.

pub fn sysid(&self) -> u32

Returns the value of sysid, or the default value if sysid is unset.

pub fn token_source(&self) -> u32

Returns the value of token_source, or the default value if token_source is unset.

pub fn admin_spoofing_user(&self) -> bool

Returns the value of admin_spoofing_user, or the default value if admin_spoofing_user is unset.

pub fn seq_num(&self) -> i32

Returns the value of seq_num, or the default value if seq_num is unset.

pub fn webapi_key_id(&self) -> u32

Returns the value of webapi_key_id, or the default value if webapi_key_id is unset.

pub fn is_from_external_source(&self) -> bool

Returns the value of is_from_external_source, or the default value if is_from_external_source is unset.

pub fn cm_sysid(&self) -> u32

Returns the value of cm_sysid, or the default value if cm_sysid is unset.

pub fn launcher_type(&self) -> u32

Returns the value of launcher_type, or the default value if launcher_type is unset.

pub fn realm(&self) -> u32

Returns the value of realm, or the default value if realm is unset.

pub fn timeout_ms(&self) -> i32

Returns the value of timeout_ms, or the default value if timeout_ms is unset.

pub fn debug_source(&self) -> &str

Returns the value of debug_source, or the default value if debug_source is unset.

pub fn debug_source_string_index(&self) -> u32

Returns the value of debug_source_string_index, or the default value if debug_source_string_index is unset.

pub fn token_id(&self) -> u64

Returns the value of token_id, or the default value if token_id is unset.

pub fn session_disposition(&self) -> ESessionDisposition

Returns the enum value of session_disposition, or the default if the field is unset or set to an invalid enum value.

pub fn set_session_disposition(&mut self, value: ESessionDisposition)

Sets session_disposition to the provided enum value.

pub fn wg_token(&self) -> &str

Returns the value of wg_token, or the default value if wg_token is unset.

pub fn webui_auth_key(&self) -> &str

Returns the value of webui_auth_key, or the default value if webui_auth_key is unset.

pub fn admin_request_spoofing_steamid(&self) -> u64

Returns the value of admin_request_spoofing_steamid, or the default value if admin_request_spoofing_steamid is unset.

pub fn is_valveds(&self) -> bool

Returns the value of is_valveds, or the default value if is_valveds is unset.

pub fn trace_tag(&self) -> u64

Returns the value of trace_tag, or the default value if trace_tag is unset.