steam::generated

Struct PublishedFileDetails

pub struct PublishedFileDetails {

    pub result: Option<u32>,
    pub publishedfileid: Option<u64>,
    pub creator: Option<u64>,
    pub creator_appid: Option<u32>,
    pub consumer_appid: Option<u32>,
    pub consumer_shortcutid: Option<u32>,
    pub filename: Option<String>,
    pub file_size: Option<u64>,
    pub preview_file_size: Option<u64>,
    pub file_url: Option<String>,
    pub preview_url: Option<String>,
    pub youtubevideoid: Option<String>,
    pub url: Option<String>,
    pub hcontent_file: Option<u64>,
    pub hcontent_preview: Option<u64>,
    pub title: Option<String>,
    pub file_description: Option<String>,
    pub short_description: Option<String>,
    pub time_created: Option<u32>,
    pub time_updated: Option<u32>,
    pub visibility: Option<u32>,
    pub flags: Option<u32>,
    pub workshop_file: Option<bool>,
    pub workshop_accepted: Option<bool>,
    pub show_subscribe_all: Option<bool>,
    pub num_comments_developer: Option<i32>,
    pub num_comments_public: Option<i32>,
    pub banned: Option<bool>,
    pub ban_reason: Option<String>,
    pub banner: Option<u64>,
    pub can_be_deleted: Option<bool>,
    pub incompatible: Option<bool>,
    pub app_name: Option<String>,
    pub file_type: Option<u32>,
    pub can_subscribe: Option<bool>,
    pub subscriptions: Option<u32>,
    pub favorited: Option<u32>,
    pub followers: Option<u32>,
    pub lifetime_subscriptions: Option<u32>,
    pub lifetime_favorited: Option<u32>,
    pub lifetime_followers: Option<u32>,
    pub lifetime_playtime: Option<u64>,
    pub lifetime_playtime_sessions: Option<u64>,
    pub views: Option<u32>,
    pub image_width: Option<u32>,
    pub image_height: Option<u32>,
    pub image_url: Option<String>,
    pub spoiler_tag: Option<bool>,
    pub shortcutid: Option<u32>,
    pub shortcutname: Option<String>,
    pub num_children: Option<u32>,
    pub num_reports: Option<u32>,
    pub previews: Vec<Preview>,
    pub tags: Vec<Tag>,
    pub children: Vec<Child>,
    pub kvtags: Vec<KvTag>,
    pub vote_data: Option<VoteData>,
    pub playtime_stats: Option<PlaytimeStats>,
    pub time_subscribed: Option<u32>,
    pub for_sale_data: Option<ForSaleData>,
    pub metadata: Option<String>,
    pub language: Option<i32>,
    pub maybe_inappropriate_sex: Option<bool>,
    pub maybe_inappropriate_violence: Option<bool>,
    pub content_descriptorids: Vec<i32>,
    pub revision_change_number: Option<u64>,
    pub revision: Option<i32>,
    pub available_revisions: Vec<i32>,
    pub reactions: Vec<Reaction>,
    pub ban_text_check_result: Option<i32>,
    pub search_score: Option<f32>,
    pub external_asset_id: Option<u64>,
    pub author_snapshots: Vec<PublishedFileAuthorSnapshot>,

}

Fields
result: Option<u32>publishedfileid: Option<u64>creator: Option<u64>creator_appid: Option<u32>consumer_appid: Option<u32>consumer_shortcutid: Option<u32>filename: Option<String>file_size: Option<u64>preview_file_size: Option<u64>file_url: Option<String>preview_url: Option<String>youtubevideoid: Option<String>url: Option<String>hcontent_file: Option<u64>hcontent_preview: Option<u64>title: Option<String>file_description: Option<String>short_description: Option<String>time_created: Option<u32>time_updated: Option<u32>visibility: Option<u32>flags: Option<u32>workshop_file: Option<bool>workshop_accepted: Option<bool>show_subscribe_all: Option<bool>num_comments_developer: Option<i32>num_comments_public: Option<i32>banned: Option<bool>ban_reason: Option<String>banner: Option<u64>can_be_deleted: Option<bool>incompatible: Option<bool>app_name: Option<String>file_type: Option<u32>can_subscribe: Option<bool>subscriptions: Option<u32>favorited: Option<u32>followers: Option<u32>lifetime_subscriptions: Option<u32>lifetime_favorited: Option<u32>lifetime_followers: Option<u32>lifetime_playtime: Option<u64>lifetime_playtime_sessions: Option<u64>views: Option<u32>image_width: Option<u32>image_height: Option<u32>image_url: Option<String>spoiler_tag: Option<bool>shortcutid: Option<u32>shortcutname: Option<String>num_children: Option<u32>num_reports: Option<u32>previews: Vec<Preview>tags: Vec<Tag>children: Vec<Child>kvtags: Vec<KvTag>vote_data: Option<VoteData>playtime_stats: Option<PlaytimeStats>time_subscribed: Option<u32>for_sale_data: Option<ForSaleData>metadata: Option<String>language: Option<i32>maybe_inappropriate_sex: Option<bool>maybe_inappropriate_violence: Option<bool>content_descriptorids: Vec<i32>revision_change_number: Option<u64>revision: Option<i32>available_revisions: Vec<i32>reactions: Vec<Reaction>ban_text_check_result: Option<i32>search_score: Option<f32>external_asset_id: Option<u64>author_snapshots: Vec<PublishedFileAuthorSnapshot>
Implementations

impl PublishedFileDetails

pub fn result(&self) -> u32

Returns the value of result, or the default value if result is unset.

pub fn publishedfileid(&self) -> u64

Returns the value of publishedfileid, or the default value if publishedfileid is unset.

pub fn creator(&self) -> u64

Returns the value of creator, or the default value if creator is unset.

pub fn creator_appid(&self) -> u32

Returns the value of creator_appid, or the default value if creator_appid is unset.

pub fn consumer_appid(&self) -> u32

Returns the value of consumer_appid, or the default value if consumer_appid is unset.

pub fn consumer_shortcutid(&self) -> u32

Returns the value of consumer_shortcutid, or the default value if consumer_shortcutid is unset.

pub fn filename(&self) -> &str

Returns the value of filename, or the default value if filename is unset.

pub fn file_size(&self) -> u64

Returns the value of file_size, or the default value if file_size is unset.

pub fn preview_file_size(&self) -> u64

Returns the value of preview_file_size, or the default value if preview_file_size is unset.

pub fn file_url(&self) -> &str

Returns the value of file_url, or the default value if file_url is unset.

pub fn preview_url(&self) -> &str

Returns the value of preview_url, or the default value if preview_url is unset.

pub fn youtubevideoid(&self) -> &str

Returns the value of youtubevideoid, or the default value if youtubevideoid is unset.

pub fn url(&self) -> &str

Returns the value of url, or the default value if url is unset.

pub fn hcontent_file(&self) -> u64

Returns the value of hcontent_file, or the default value if hcontent_file is unset.

pub fn hcontent_preview(&self) -> u64

Returns the value of hcontent_preview, or the default value if hcontent_preview is unset.

pub fn title(&self) -> &str

Returns the value of title, or the default value if title is unset.

pub fn file_description(&self) -> &str

Returns the value of file_description, or the default value if file_description is unset.

pub fn short_description(&self) -> &str

Returns the value of short_description, or the default value if short_description is unset.

pub fn time_created(&self) -> u32

Returns the value of time_created, or the default value if time_created is unset.

pub fn time_updated(&self) -> u32

Returns the value of time_updated, or the default value if time_updated is unset.

pub fn visibility(&self) -> u32

Returns the value of visibility, or the default value if visibility is unset.

pub fn flags(&self) -> u32

Returns the value of flags, or the default value if flags is unset.

pub fn workshop_file(&self) -> bool

Returns the value of workshop_file, or the default value if workshop_file is unset.

pub fn workshop_accepted(&self) -> bool

Returns the value of workshop_accepted, or the default value if workshop_accepted is unset.

pub fn show_subscribe_all(&self) -> bool

Returns the value of show_subscribe_all, or the default value if show_subscribe_all is unset.

pub fn num_comments_developer(&self) -> i32

Returns the value of num_comments_developer, or the default value if num_comments_developer is unset.

pub fn num_comments_public(&self) -> i32

Returns the value of num_comments_public, or the default value if num_comments_public is unset.

pub fn banned(&self) -> bool

Returns the value of banned, or the default value if banned is unset.

pub fn ban_reason(&self) -> &str

Returns the value of ban_reason, or the default value if ban_reason is unset.

pub fn banner(&self) -> u64

Returns the value of banner, or the default value if banner is unset.

pub fn can_be_deleted(&self) -> bool

Returns the value of can_be_deleted, or the default value if can_be_deleted is unset.

pub fn incompatible(&self) -> bool

Returns the value of incompatible, or the default value if incompatible is unset.

pub fn app_name(&self) -> &str

Returns the value of app_name, or the default value if app_name is unset.

pub fn file_type(&self) -> u32

Returns the value of file_type, or the default value if file_type is unset.

pub fn can_subscribe(&self) -> bool

Returns the value of can_subscribe, or the default value if can_subscribe is unset.

pub fn subscriptions(&self) -> u32

Returns the value of subscriptions, or the default value if subscriptions is unset.

pub fn favorited(&self) -> u32

Returns the value of favorited, or the default value if favorited is unset.

pub fn followers(&self) -> u32

Returns the value of followers, or the default value if followers is unset.

pub fn lifetime_subscriptions(&self) -> u32

Returns the value of lifetime_subscriptions, or the default value if lifetime_subscriptions is unset.

pub fn lifetime_favorited(&self) -> u32

Returns the value of lifetime_favorited, or the default value if lifetime_favorited is unset.

pub fn lifetime_followers(&self) -> u32

Returns the value of lifetime_followers, or the default value if lifetime_followers is unset.

pub fn views(&self) -> u32

Returns the value of views, or the default value if views is unset.

pub fn image_width(&self) -> u32

Returns the value of image_width, or the default value if image_width is unset.

pub fn image_height(&self) -> u32

Returns the value of image_height, or the default value if image_height is unset.

pub fn image_url(&self) -> &str

Returns the value of image_url, or the default value if image_url is unset.

pub fn spoiler_tag(&self) -> bool

Returns the value of spoiler_tag, or the default value if spoiler_tag is unset.

pub fn shortcutid(&self) -> u32

Returns the value of shortcutid, or the default value if shortcutid is unset.

pub fn shortcutname(&self) -> &str

Returns the value of shortcutname, or the default value if shortcutname is unset.

pub fn num_children(&self) -> u32

Returns the value of num_children, or the default value if num_children is unset.

pub fn num_reports(&self) -> u32

Returns the value of num_reports, or the default value if num_reports is unset.

pub fn time_subscribed(&self) -> u32

Returns the value of time_subscribed, or the default value if time_subscribed is unset.

pub fn metadata(&self) -> &str

Returns the value of metadata, or the default value if metadata is unset.

pub fn language(&self) -> i32

Returns the value of language, or the default value if language is unset.

pub fn lifetime_playtime(&self) -> u64

Returns the value of lifetime_playtime, or the default value if lifetime_playtime is unset.

pub fn lifetime_playtime_sessions(&self) -> u64

Returns the value of lifetime_playtime_sessions, or the default value if lifetime_playtime_sessions is unset.

pub fn maybe_inappropriate_sex(&self) -> bool

Returns the value of maybe_inappropriate_sex, or the default value if maybe_inappropriate_sex is unset.

pub fn maybe_inappropriate_violence(&self) -> bool

Returns the value of maybe_inappropriate_violence, or the default value if maybe_inappropriate_violence is unset.

pub fn revision_change_number(&self) -> u64

Returns the value of revision_change_number, or the default value if revision_change_number is unset.

pub fn revision(&self) -> EPublishedFileRevision

Returns the enum value of revision, or the default if the field is unset or set to an invalid enum value.

pub fn set_revision(&mut self, value: EPublishedFileRevision)

Sets revision to the provided enum value.

pub fn available_revisions( &self, ) -> FilterMap<Cloned<Iter<'_, i32>>, fn(i32) -> Option<EPublishedFileRevision>>

Returns an iterator which yields the valid enum values contained in available_revisions.

pub fn push_available_revisions(&mut self, value: EPublishedFileRevision)

Appends the provided enum value to available_revisions.

pub fn ban_text_check_result(&self) -> EBanContentCheckResult

Returns the enum value of ban_text_check_result, or the default if the field is unset or set to an invalid enum value.

pub fn set_ban_text_check_result(&mut self, value: EBanContentCheckResult)

Sets ban_text_check_result to the provided enum value.

pub fn content_descriptorids( &self, ) -> FilterMap<Cloned<Iter<'_, i32>>, fn(i32) -> Option<EContentDescriptorId>>

Returns an iterator which yields the valid enum values contained in content_descriptorids.

pub fn push_content_descriptorids(&mut self, value: EContentDescriptorId)

Appends the provided enum value to content_descriptorids.

pub fn search_score(&self) -> f32

Returns the value of search_score, or the default value if search_score is unset.

pub fn external_asset_id(&self) -> u64

Returns the value of external_asset_id, or the default value if external_asset_id is unset.