steam::generated

Struct UserSystemInformation

pub struct UserSystemInformation {

    pub manufacturer: Option<String>,
    pub model: Option<String>,
    pub dx_video_card: Option<String>,
    pub dx_vendorid: Option<i32>,
    pub dx_deviceid: Option<i32>,
    pub num_gpu: Option<u32>,
    pub system_ram: Option<u64>,
    pub os: Option<String>,
    pub cpu_vendor: Option<String>,
    pub cpu_name: Option<String>,
    pub gaming_device_type: Option<u32>,
    pub dx_driver_version: Option<String>,
    pub adapter_description: Option<String>,
    pub driver_version: Option<String>,
    pub driver_date: Option<String>,
    pub vram_size: Option<u32>,
    pub screen_width: Option<u32>,
    pub screen_height: Option<u32>,

}

Fields
manufacturer: Option<String>model: Option<String>dx_video_card: Option<String>dx_vendorid: Option<i32>dx_deviceid: Option<i32>num_gpu: Option<u32>system_ram: Option<u64>os: Option<String>cpu_vendor: Option<String>cpu_name: Option<String>gaming_device_type: Option<u32>dx_driver_version: Option<String>adapter_description: Option<String>driver_version: Option<String>driver_date: Option<String>vram_size: Option<u32>screen_width: Option<u32>screen_height: Option<u32>
Implementations

impl UserSystemInformation

pub fn manufacturer(&self) -> &str

Returns the value of manufacturer, or the default value if manufacturer is unset.

pub fn model(&self) -> &str

Returns the value of model, or the default value if model is unset.

pub fn dx_video_card(&self) -> &str

Returns the value of dx_video_card, or the default value if dx_video_card is unset.

pub fn dx_vendorid(&self) -> i32

Returns the value of dx_vendorid, or the default value if dx_vendorid is unset.

pub fn dx_deviceid(&self) -> i32

Returns the value of dx_deviceid, or the default value if dx_deviceid is unset.

pub fn num_gpu(&self) -> u32

Returns the value of num_gpu, or the default value if num_gpu is unset.

pub fn system_ram(&self) -> u64

Returns the value of system_ram, or the default value if system_ram is unset.

pub fn os(&self) -> &str

Returns the value of os, or the default value if os is unset.

pub fn cpu_vendor(&self) -> &str

Returns the value of cpu_vendor, or the default value if cpu_vendor is unset.

pub fn cpu_name(&self) -> &str

Returns the value of cpu_name, or the default value if cpu_name is unset.

pub fn gaming_device_type(&self) -> u32

Returns the value of gaming_device_type, or the default value if gaming_device_type is unset.

pub fn dx_driver_version(&self) -> &str

Returns the value of dx_driver_version, or the default value if dx_driver_version is unset.

pub fn adapter_description(&self) -> &str

Returns the value of adapter_description, or the default value if adapter_description is unset.

pub fn driver_version(&self) -> &str

Returns the value of driver_version, or the default value if driver_version is unset.

pub fn driver_date(&self) -> &str

Returns the value of driver_date, or the default value if driver_date is unset.

pub fn vram_size(&self) -> u32

Returns the value of vram_size, or the default value if vram_size is unset.

pub fn screen_width(&self) -> u32

Returns the value of screen_width, or the default value if screen_width is unset.

pub fn screen_height(&self) -> u32

Returns the value of screen_height, or the default value if screen_height is unset.